+++ lib/shrine/plugins/activerecord.rb
@@ -32,17 +32,23 @@ class Shrine
 
           if shrine_class.opts[:activerecord][:callbacks]
             model.before_save do
-              send(:"#{name}_attacher").send(:activerecord_before_save)
+              if send(:"#{name}_attacher").changed?
+                send(:"#{name}_attacher").send(:activerecord_before_save)
+              end
             end
 
             [:create, :update].each do |action|
               model.after_commit on: action do
-                send(:"#{name}_attacher").send(:activerecord_after_save)
+                if send(:"#{name}_attacher").changed?
+                  send(:"#{name}_attacher").send(:activerecord_after_save)
+                end
               end
             end
 
             model.after_commit on: :destroy do
-              send(:"#{name}_attacher").send(:activerecord_after_destroy)
+              if send(:"#{name}_attacher").attached?
+                send(:"#{name}_attacher").send(:activerecord_after_destroy)
+              end
             end
           end
 
@@ -76,15 +82,11 @@ class Shrine
 
         # Calls Attacher#save. Called before model save.
         def activerecord_before_save
-          return unless changed?
-
           save
         end
 
         # Finalizes attachment and persists changes. Called after model save.
         def activerecord_after_save
-          return unless changed?
-
           finalize
           persist
         end